Move the ACPI tables code into a separate file
